requirements.txt
Flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
scikit-learn==1.3.0
transformers==4.33.0
torch==2.0.1
Pillow==10.0.0
opencv-python==4.8.0.76
gunicorn==21.2.0
